react/backend/requirements.txt
flask>=2.0.1
flask-cors>=3.0.10
numpy>=1.26.0
opencv-python>=4.5.3
mediapipe>=0.10.9
joblib>=1.3.2
pillow>=8.3.2
python-dotenv>=0.19.0 